forms: phase 1.2e — extract ListBox/ComboBox item natives
Continue moving per-kind natives. The ListBox + ComboBox item-management
cluster leaves forms_module.c as a single TU shared by both kinds:

  src/controls/ctl_listbox.{c,h}
    - native_add_item            (LB_ADDSTRING / CB_ADDSTRING)
    - native_clear_items         (LB_RESETCONTENT / CB_RESETCONTENT)
    - native_get_item_count      (LB_GETCOUNT / CB_GETCOUNT)
    - native_get_item            (LB_GETTEXT / CB_GETLBTEXT)
    - native_get_items           (loop over count, returns array of
                                   strings)
    - native_remove_item         (LB_DELETESTRING / CB_DELETESTRING)
    - native_get_selected_index  (LB_GETCURSEL / CB_GETCURSEL)
    - native_set_selected_index  (LB_SETCURSEL / CB_SETCURSEL)
    - list_item_text helper      (shared by getItem and getItems)

Each native dispatches on s->kind to the matching Win32 control
class.  Phase 1.1 attached these methods only to forms_listbox /
forms_combobox so the runtime kind switch is now belt-and-braces;
the per-kind meta tables already gate at the dispatcher level.

forms_module.c #includes ctl_listbox.h in place of the in-place
definitions.

Main TU shrinks from 4346 → 4178 lines (cumulative -1059 lines from
the 5237-line baseline).

All 100 C unit tests still pass.
